# server/src/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the scrumlr backend server.",
        "title": "Scrumlr backend",
        "termsOfService": "https://scrumlr.io/terms",
        "contact": {
            "email": "info@scrumlr.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/inovex/scrumlr.io/blob/main/LICENSE"
        },
        "version": "5.4.0"
    },
    "paths": {
...
        "/templates/{boardid}/columns": {
            "get": {
                "description": "Get all column templates from a board template",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "column templates"
                ],
                "summary": "Get all column templates from a board template",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token to authenticate",
                        "name": "Cookie",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "id of the board template",
                        "name": "boardid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/columntemplates.ColumnTemplate"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/common.APIError"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests"
                    }
                }
            },
            "post": {
                "description": "Create a column template for an existing board template",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "column templates"
                ],
                "summary": "Create a column template for a board template",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token to authenticate",
                        "name": "Cookie",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "id of the board template",
                        "name": "boardid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "column template to create",
                        "name": "columntemplate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/columntemplates.ColumnTemplateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/columntemplates.ColumnTemplate"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.APIError"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests"
                    }
                }
            }
        },
        "/templates/{boardid}/columns/{id}": {
            "get": {
                "description": "Get a column template from a board template",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "column templates"
                ],
                "summary": "Get a column template from a board template",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token to authenticate",
                        "name": "Cookie",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "id of the board template",
                        "name": "boardid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "id of the column template",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/columntemplates.ColumnTemplate"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/common.APIError"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests"
                    }
                }
            },
            "put": {
                "description": "Update a column template from a board template",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "column templates"
                ],
                "summary": "Update a column template from a board template",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token to authenticate",
                        "name": "Cookie",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "id of the board template",
                        "name": "boardid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "id of the column template to update",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/columntemplates.ColumnTemplate"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/common.APIError"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests"
                    }
                }
            },
            "delete": {
                "description": "Delete a column template from a board template",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "column templates"
                ],
                "summary": "Delete a column template from a board template",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token to authenticate",
                        "name": "Cookie",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "id of the board template",
                        "name": "boardid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "id of the column template to delete",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/common.APIError"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "columntemplates.ColumnTemplate": {
            "type": "object",
            "properties": {
                "color": {
                    "description": "The column template color.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/common.Color"
                        }
                    ]
                },
                "description": {
                    "description": "The description of a board template column.",
                    "type": "string"
                },
                "id": {
                    "description": "The column template id.",
                    "type": "string"
                },
                "index": {
                    "description": "The column template rank.",
                    "type": "integer"
                },
                "name": {
                    "description": "The column template name.",
                    "type": "string"
                },
                "template": {
                    "description": "The board template id, that this column template is bound to.",
                    "type": "string"
                },
                "visible": {
                    "description": "The column template visibility.",
                    "type": "boolean"
                }
            }
        },
        "columntemplates.ColumnTemplateRequest": {
            "type": "object",
            "properties": {
                "color": {
                    "description": "The column template color to set.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/common.Color"
                        }
                    ]
                },
                "description": {
                    "description": "The columnTemplate description to set.",
                    "type": "string"
                },
                "index": {
                    "description": "Sets the index of this column template in the sort order.",
                    "type": "integer"
                },
                "name": {
                    "description": "The column template name to set.",
                    "type": "string"
                },
                "visible": {
                    "description": "Sets whether this column template should be visible to regular participants.\nThe default value on creation is 'false'.",
                    "type": "boolean"
                }
            }
        },
        "common.APIError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "common.Color": {
            "type": "string",
            "enum": [
                "backlog-blue",
                "goal-green",
                "value-violet",
                "online-orange",
                "planning-pink",
                "poker-purple",
                "yielding-yellow"
            ],
            "x-enum-varnames": [
                "ColorBacklogBlue",
                "ColorGoalGreen",
                "ColorValueViolet",
                "ColorOnlineOrange",
                "ColorPlanningPink",
                "ColorPokerPurple",
                "ColorYieldingYellow"
            ]
        },
...
    }
}